move cuda_graph from fluid to phi
