+++ paddle/fluid/platform/device/gpu/cuda/CMakeLists.txt
@@ -1,8 +1,4 @@
 nv_library(
-  cuda_graph
-  SRCS cuda_graph.cc
-  DEPS enforce)
-nv_library(
   cuda_profiler
   SRCS cuda_profiler.cc
   DEPS enforce)
