--- paddle/phi/backends/xpu/CMakeLists.txt
@@ -1,4 +1,0 @@
-cc_library(
-  phi_xpu_op_list
-  SRCS xpu_op_list.cc
-  DEPS glog)
